requirements.txt
WTForms~=3.0.1
Flask~=2.0.2
Werkzeug~=2.0.2
python-dotenv~=0.19.2
SQLAlchemy~=1.4.31
gunicorn~=20.1.0
Flask-Bootstrap~=3.3.7.1
Flask-CKEditor~=0.4.6
Flask-Gravatar~=0.5.0
Flask-Login~=0.5.0
Flask-WTF~=1.0.0
Flask-SQLAlchemy~=2.5.1
email-validator~=1.1.3
psycopg2-binary~=2.9.3